dikstruiyua

+++ src/main/org/insa/algo/utils/Label.java
@@ -7,17 +7,17 @@ public class Label {
 
 	private Node sommet_courant;
 	private boolean marque;
-	private int cout;
+	private double cout;
 	private Arc pere;
 	
-	public Label (Node courant, Arc papa, boolean mark, int cost) {
+	public Label (Node courant, Arc papa, boolean mark, double cost) {
 		this.sommet_courant = courant ;
 		this.marque = mark ;
 		this.cout = cost ;
 		this.pere = papa ;
 	}
 	
-	int getcost() {
+	double getcost() {
 		return this.cout;
 	}
 }
